+++ README.md
@@ -1,2 +1,19 @@
-# Virtual-Keyboard
-Virtual Keyboard
+Vanilla JS project: **Virtual Keyboard**.
+
+#### Project deploy link [Virtual-Keyboard](https://github.com/Yan-Black/Virtual-Keyboard/tree/gh-pages).
+
+## Implemented features
+* Сompletely repeats the functionality of the physical keyboard
+* Language changing
+* Saving the selected language to the localStorage
+* Correct work of CapsLock and Shift keys
+
+### Availible Scripts
+* npm install (install all necessary dependencies)
+* npm run start (starts project at local server)
+* npm run build (builds a production version)
+* npm run lint (runs eslint at src/ folder)
+* npm run lint:fix (fix all fixabled problems)
+
+### Code quality
+* Eslint (Air-bnb style guide)
