+++ package.json
@@ -3,12 +3,28 @@
   "version": "1.0.0",
   "description": "Virtual Keyboard",
   "main": "script.js",
-  "dependencies": {},
+  "dependencies": {
+    "html-webpack-plugin": "^4.4.1"
+  },
   "devDependencies": {
-    "eslint": "^6.8.0"
+    "@babel/core": "^7.11.6",
+    "@babel/preset-env": "^7.11.5",
+    "@babel/preset-react": "^7.10.4",
+    "babel-loader": "^8.1.0",
+    "css-loader": "^4.2.2",
+    "eslint": "^7.8.1",
+    "eslint-config-airbnb-base": "^14.2.0",
+    "eslint-plugin-import": "^2.22.0",
+    "style-loader": "^1.2.1",
+    "webpack": "^4.44.1",
+    "webpack-cli": "^3.3.12",
+    "webpack-dev-server": "^3.11.0"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "webpack-dev-server --mode development",
+    "build": "webpack --mode production",
+    "lint": "eslint src/",
+    "lint:fix": "eslint --fix src/"
   },
   "repository": {
     "type": "git",
