+++ public/index.html
@@ -3,10 +3,9 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="style.css">
     <title>Virtual Keyboard</title>
 </head>
 <body>
-    <script src="./dist/bundle.js"></script>
+
 </body>
 </html>
