+++ src/components/Button.js
@@ -1,4 +1,4 @@
-export class Button {
+export default class Button {
   constructor(type) {
     this.type = type;
   }
